fix(ET-1633): Change primary_at from u64 to f64

See merge request rust/proton-rust!291

+++ core/core-common/src/db/migrations/account/v0.rs
@@ -44,7 +44,7 @@ async fn create_table_core_accounts(tx: &Bond<'_>) -> Result<(), StashError> {
                 primary_addr TEXT,
 
                 -- Timestamp of when account was made primary
-                primary_at INTEGER
+                primary_at REAL
             )
         ",
         vec![],
